+++ lib/freshbooks_api_client/schema/time_entry.ex
@@ -5,8 +5,7 @@ defmodule FreshbooksApiClient.Schema.TimeEntry do
   It uses a FreshbooksApiClient.Schema
   """
 
-  use FreshbooksApiClient.Schema, resource: "time_entry",
-    resources: "time_entries"
+  use FreshbooksApiClient.Schema
 
   api_schema do
     field :time_entry_id, :integer
